Add files via upload
This is code to look at the orientation of a TC on a plan view map, with wind barbs, topography, and pressure contours. I made it to view the location of a cross section I took relative to the location of Ivan's winds at that time(important for orographic enhancement). Animation code included for the plots the script creates
I am planning to update this sometime soon with locations of multiple cross sections along a given stretch, which I will average (average cross section, reduce noise).