Add files via upload
Calculate and plot hourly rainrate for multiple timesteps. Comparison of rainrate for 2 different model runs included if desired/needed

Note: Rainrate is calculated in such a way that the rainrate data at a given time is ACTUALLY the rainrate for the previous time index.
For example, if 09/17/04 at 03Z is index 7 in a list of times, the rainrate for this hour would actually be index 8, so plot index 8 for the RR in the hour of 03Z. (Index 7 would actually be the RR for hour 09/17 at 02Z)

Note 2: The rainrates calculated are ONLY hourly IF your output files are hourly. If you have 3-hourly output, then this calculates 3-hourly rainrates